mpi/src: size of a data is stored in the starpu-mpi request as a size_t

+++ mpi/src/starpu_mpi_datatype.c
@@ -127,7 +127,7 @@ static handle_to_datatype_func handle_to_datatype_funcs[STARPU_MAX_INTERFACE_ID]
 	[STARPU_MULTIFORMAT_INTERFACE_ID] = NULL,
 };
 
-int starpu_mpi_handle_to_datatype(starpu_data_handle_t data_handle, MPI_Datatype *datatype, int *count)
+int starpu_mpi_handle_to_datatype(starpu_data_handle_t data_handle, MPI_Datatype *datatype, size_t *count)
 {
 	enum starpu_data_interface_id id = starpu_handle_get_interface_id(data_handle);
